+++ CMakeLists.txt
@@ -18,13 +18,14 @@ include(cmake/FindCompilerFlag.cmake)
 
 option(ENABLE_RTTI "Enable compiling with real time type information" OFF)
 option(ENABLE_LLVM_SHARED "Enable linking LLVM as a shared library" OFF)
+option(LIBBPF_ONLY "Build JUST the libbpf library and related headers" OFF)
 option(ENABLE_CLANG_JIT "Enable Loading BPF through Clang Frontend" ON)
 option(ENABLE_USDT "Enable User-level Statically Defined Tracing" ON)
 CMAKE_DEPENDENT_OPTION(ENABLE_CPP_API "Enable C++ API" ON "ENABLE_USDT" OFF)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
 
-if(NOT PYTHON_ONLY AND ENABLE_CLANG_JIT)
+if(NOT PYTHON_ONLY AND NOT LIBBPF_ONLY)
 find_package(BISON)
 find_package(FLEX)
 find_package(LLVM REQUIRED CONFIG)
@@ -80,16 +81,16 @@ endif()
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_STANDARD 11)
 
-endif(NOT PYTHON_ONLY AND ENABLE_CLANG_JIT)
+endif(NOT PYTHON_ONLY AND NOT LIBBPF_ONLY)
 
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall ${CXX_ISYSTEM_DIRS}")
 
 add_subdirectory(src)
 add_subdirectory(introspection)
-if(ENABLE_CLANG_JIT)
+if(NOT LIBBPF_ONLY)
 add_subdirectory(examples)
 add_subdirectory(man)
 add_subdirectory(tests)
 add_subdirectory(tools)
-endif(ENABLE_CLANG_JIT)
+endif(NOT LIBBPF_ONLY)
