+++ src/CMakeLists.txt
@@ -15,7 +15,7 @@ include_directories(${CMAKE_CURRENT_BINARY_DIR})
 if(NOT PYTHON_ONLY)
 add_subdirectory(cc)
 endif()
-if(ENABLE_CLANG_JIT)
+if(NOT LIBBPF_ONLY)
 add_subdirectory(python)
 add_subdirectory(lua)
 endif()
