+++ src/cc/frontends/clang/loader.h
@@ -21,6 +21,7 @@
 #include <string>
 
 #include "table_storage.h"
+#include "func_source.h"
 
 namespace llvm {
 class Module;
@@ -29,23 +30,6 @@ class MemoryBuffer;
 }
 
 namespace ebpf {
-
-class FuncSource {
-  class SourceCode {
-   public:
-    SourceCode(const std::string& s1 = "", const std::string& s2 = ""): src_(s1), src_rewritten_(s2) {}
-    std::string src_;
-    std::string src_rewritten_;
-  };
-  std::map<std::string, SourceCode> funcs_;
- public:
-  FuncSource() {}
-  void clear() { funcs_.clear(); }
-  const char * src(const std::string& name);
-  const char * src_rewritten(const std::string& name);
-  void set_src(const std::string& name, const std::string& src);
-  void set_src_rewritten(const std::string& name, const std::string& src);
-};
 
 class ClangLoader {
  public:
