+++ tests/CMakeLists.txt
@@ -7,7 +7,7 @@ set(TEST_WRAPPER ${CMAKE_CURRENT_BINARY_DIR}/wrapper.sh)
 add_test(NAME style-check COMMAND ${CMAKE_SOURCE_DIR}/scripts/style-check.sh)
 set_tests_properties(style-check PROPERTIES PASS_REGULAR_EXPRESSION ".*")
 
-if(ENABLE_CLANG_JIT)
+if(NOT LIBBPF_ONLY)
 add_subdirectory(cc)
 add_subdirectory(python)
 add_subdirectory(lua)
